projects/agentic_portfolio/pyproject.toml
[project]
name = "agentic-portfolio"
version = "0.1.0"
description = "Phase 1 foundation for the agentic portfolio backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.116.0",
  "pydantic-settings>=2.10.0",
  "uvicorn>=0.35.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28.0",
  "pytest>=8.4.0",
]

[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
exclude = ["workspace*", "tests*"]
